+++ cmd/carbonapi/config/config.go
@@ -60,6 +60,8 @@ type ConfigType struct {
 	DefaultColors              map[string]string  `mapstructure:"defaultColors"`
 	GraphTemplates             string             `mapstructure:"graphTemplates"`
 	FunctionsConfigs           map[string]string  `mapstructure:"functionsConfig"`
+	HeadersToPass              []string           `mapstructure:"headersToPass"`
+	// HeadersToLog               []string           `mapstructure:"headersToLog"`
 
 	QueryCache cache.BytesCache `mapstructure:"-" json:"-"`
 	FindCache  cache.BytesCache `mapstructure:"-" json:"-"`