+++ cmd/carbonapi/http/enrichcontext.go
@@ -1,0 +1,24 @@
+package http
+
+import (
+	"net/http"
+
+	utilctx "github.com/go-graphite/carbonapi/util/ctx"
+)
+
+// TrackConnections exports via expvar a list of all currently executing requests
+func enrichContextWithHeaders(headersToPass []string, fn http.HandlerFunc) http.HandlerFunc {
+	return func(w http.ResponseWriter, req *http.Request) {
+		headersToPassMap := make(map[string]string)
+		for _, name := range headersToPass {
+			h := req.Header.Get(name)
+			if h != "" {
+				headersToPassMap[name] = h
+			}
+		}
+		ctx := utilctx.SetPassHeaders(req.Context(), headersToPassMap)
+		req = req.WithContext(ctx)
+
+		fn(w, req)
+	}
+}