+++ util/ctx/ctx.go
@@ -11,7 +11,8 @@ const (
 	HeaderUUIDAPI    = "X-CTX-CarbonAPI-UUID"
 	HeaderUUIDZipper = "X-CTX-CarbonZipper-UUID"
 
-	uuidKey key = 0
+	uuidKey key = iota
+	headersKey
 )
 
 func ifaceToString(v interface{}) string {
@@ -23,6 +24,26 @@ func ifaceToString(v interface{}) string {
 
 func getCtxString(ctx context.Context, k key) string {
 	return ifaceToString(ctx.Value(k))
+}
+
+func getCtxMapString(ctx context.Context, k key) map[string]string {
+	v := ctx.Value(k)
+	if v != nil {
+		vv, ok := v.(map[string]string)
+		if !ok {
+			return map[string]string{}
+		}
+		return vv
+	}
+	return map[string]string{}
+}
+
+func GetPassHeaders(ctx context.Context) map[string]string {
+	return getCtxMapString(ctx, headersKey)
+}
+
+func SetPassHeaders(ctx context.Context, h map[string]string) context.Context {
+	return context.WithValue(ctx, headersKey, h)
 }
 
 func GetUUID(ctx context.Context) string {
@@ -42,6 +63,15 @@ func ParseCtx(h http.HandlerFunc, uuidKey string) http.HandlerFunc {
 
 		h.ServeHTTP(rw, req.WithContext(ctx))
 	})
+}
+
+func MarshalPassHeaders(ctx context.Context, response *http.Request) *http.Request {
+	headers := GetPassHeaders(ctx)
+	for name, value := range headers {
+		response.Header.Add(name, value)
+	}
+
+	return response
 }
 
 func MarshalCtx(ctx context.Context, response *http.Request, uuidKey string) *http.Request {